fix: add delta time to activated arcane symbol rotation (#105)

+++ MechoSoma/RUNTIME/arcane_menu.h
@@ -225,8 +225,8 @@ struct mchArcaneScreenElement
 
 	float Timer;
 
-	int clockTimer;
-	int clockPeriod;
+	float clockTimer;
+	float clockPeriod;
 
 	Vect3f R;
 	Vect3f R0;
